using half-precision Adam statistics

+++ src/tests/clip_core_unit_tests.py
@@ -143,7 +143,7 @@ class CLIPGPUUnitTest(unittest.TestCase):
 
         with autocast():
             # Get cosine similarities
-            logits = clip_model(imgs.half(), tokenized_words.half())
+            logits = clip_model(imgs, tokenized_words)
             # Compute loss and backpropagation
             loss = loss_function(logits)
